Refactor common functionalities as extension methods
