+++ android/anc/src/main/java/org/smartregister/fhircore/anc/data/AncPatientRepository.kt
@@ -27,13 +27,13 @@ import org.smartregister.fhircore.anc.AncApplication
 import org.smartregister.fhircore.anc.data.model.AncPatientDetailItem
 import org.smartregister.fhircore.anc.data.model.AncPatientItem
 import org.smartregister.fhircore.anc.data.model.CarePlanItem
-import org.smartregister.fhircore.anc.ui.anccare.details.AncPatientItemMapper.extractAge
-import org.smartregister.fhircore.anc.ui.anccare.details.AncPatientItemMapper.extractGender
-import org.smartregister.fhircore.anc.ui.anccare.details.AncPatientItemMapper.extractName
 import org.smartregister.fhircore.engine.data.domain.util.DomainMapper
 import org.smartregister.fhircore.engine.data.domain.util.RegisterRepository
 import org.smartregister.fhircore.engine.util.DefaultDispatcherProvider
 import org.smartregister.fhircore.engine.util.DispatcherProvider
+import org.smartregister.fhircore.engine.util.extension.extractAge
+import org.smartregister.fhircore.engine.util.extension.extractGender
+import org.smartregister.fhircore.engine.util.extension.extractName
 import org.smartregister.fhircore.engine.util.extension.searchPatients
 
 class AncPatientRepository(
