+++ android/engine/src/main/res/values/strings.xml
@@ -35,4 +35,8 @@
     <string name="str_next">Next</string>
     <string name="str_previous">Previous</string>
     <string name="str_page_info">Page %1$d of %2$d</string>
+    <string name="male">Male</string>
+    <string name="female">Female</string>
+    <string name="other">Other</string>
+    <string name="unknown">Unknown</string>
 </resources>
